# internal/oas/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

// Ref: #/components/schemas/Car
type Car struct {
	ID           int    `json:"id"`
	Manufacturer string `json:"manufacturer"`
	Model        string `json:"model"`
	Year         int    `json:"year"`
	Color        string `json:"color"`
}

// GetID returns the value of ID.
func (s *Car) GetID() int {
	return s.ID
}

// GetManufacturer returns the value of Manufacturer.
func (s *Car) GetManufacturer() string {
	return s.Manufacturer
}

// GetModel returns the value of Model.
func (s *Car) GetModel() string {
	return s.Model
}

// GetYear returns the value of Year.
func (s *Car) GetYear() int {
	return s.Year
}

// GetColor returns the value of Color.
func (s *Car) GetColor() string {
	return s.Color
}

// SetID sets the value of ID.
func (s *Car) SetID(val int) {
	s.ID = val
}

// SetManufacturer sets the value of Manufacturer.
func (s *Car) SetManufacturer(val string) {
	s.Manufacturer = val
}

// SetModel sets the value of Model.
func (s *Car) SetModel(val string) {
	s.Model = val
}

// SetYear sets the value of Year.
func (s *Car) SetYear(val int) {
	s.Year = val
}

// SetColor sets the value of Color.
func (s *Car) SetColor(val string) {
	s.Color = val
}

func (*Car) getCarByIdRes()    {}
func (*Car) updateCarByIdRes() {}

// DeleteCarByIdNoContent is response for DeleteCarById operation.
type DeleteCarByIdNoContent struct{}

func (*DeleteCarByIdNoContent) deleteCarByIdRes() {}

// DeleteCarByIdNotFound is response for DeleteCarById operation.
type DeleteCarByIdNotFound struct{}

func (*DeleteCarByIdNotFound) deleteCarByIdRes() {}

// GetCarByIdNotFound is response for GetCarById operation.
type GetCarByIdNotFound struct{}

func (*GetCarByIdNotFound) getCarByIdRes() {}

// Ref: #/components/schemas/NewCar
type NewCar struct {
	Manufacturer string `json:"manufacturer"`
	Model        string `json:"model"`
	Year         int    `json:"year"`
	Color        string `json:"color"`
}

// GetManufacturer returns the value of Manufacturer.
func (s *NewCar) GetManufacturer() string {
	return s.Manufacturer
}

// GetModel returns the value of Model.
func (s *NewCar) GetModel() string {
	return s.Model
}

// GetYear returns the value of Year.
func (s *NewCar) GetYear() int {
	return s.Year
}

// GetColor returns the value of Color.
func (s *NewCar) GetColor() string {
	return s.Color
}

// SetManufacturer sets the value of Manufacturer.
func (s *NewCar) SetManufacturer(val string) {
	s.Manufacturer = val
}

// SetModel sets the value of Model.
func (s *NewCar) SetModel(val string) {
	s.Model = val
}

// SetYear sets the value of Year.
func (s *NewCar) SetYear(val int) {
	s.Year = val
}

// SetColor sets the value of Color.
func (s *NewCar) SetColor(val string) {
	s.Color = val
}

// UpdateCarByIdNotFound is response for UpdateCarById operation.
type UpdateCarByIdNotFound struct{}

func (*UpdateCarByIdNotFound) updateCarByIdRes() {}